fix: 修复购物车点击商品事件和步进器事件冲突的问题，修复步进器为0时候的弹窗点击取消还会删除商品的bug

+++ src/pages/cart/cart.js
@@ -81,17 +81,19 @@ Page({
       confirmColor: 'red',
       success: (res) => {
         console.log(res);
-        const params = {
-          carId
-        }
-        deleteShoppingCart(params).then(res => {
-          console.log(res);
-          wx.showToast({
-            title: '删除成功',
-            duration: 500
+        if (res.confirm) {
+          const params = {
+            carId
+          }
+          deleteShoppingCart(params).then(res => {
+            console.log(res);
+            wx.showToast({
+              title: '删除成功',
+              duration: 500
+            })
+            this.getShoppingCartData();
           })
-          this.getShoppingCartData();
-        })
+        }
       },
       fail: (err) => {
         console.log(err);
